Renamed package to "remote" from "server".
