+++ gremlin-core/src/main/java/org/apache/tinkerpop/gremlin/process/remote/ServerGraph.java
@@ -16,11 +16,11 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.tinkerpop.gremlin.process.server;
+package org.apache.tinkerpop.gremlin.process.remote;
 
 import org.apache.commons.configuration.Configuration;
 import org.apache.tinkerpop.gremlin.process.computer.GraphComputer;
-import org.apache.tinkerpop.gremlin.process.server.traversal.strategy.ServerStrategy;
+import org.apache.tinkerpop.gremlin.process.remote.traversal.strategy.ServerStrategy;
 import org.apache.tinkerpop.gremlin.process.traversal.TraversalStrategies;
 import org.apache.tinkerpop.gremlin.structure.Edge;
 import org.apache.tinkerpop.gremlin.structure.Graph;
